refactor(platform-wallet): extract send_transaction to its own file
Move send_transaction, select_utxos_for_payment, sign_transaction_inputs,
and related helpers (estimate_standard_tx_size, MIN_ASSET_LOCK_FEE,
DUST_THRESHOLD) from wallet.rs to core/send_transaction.rs.
