+++ packages/rs-platform-wallet/src/wallet/core/mod.rs
@@ -1,5 +1,6 @@
 pub mod balance;
 pub mod balance_handler;
+mod send_transaction;
 pub mod wallet;
 
 pub use balance::WalletBalance;
